[FRONTEND] Move AddProductForm
